Start event plugin skeleton
This adds a basic stub of the main event plugin with a few
endpoints for subscription CRUD.

It adds in enough logic to be able to register it as a builtin plugin
and mount it, but not much else.

This is based heavily on `builtin/logical/database`, though
with a lot of things removed that we won't need or aren't ready for.